Vectorization/requirements.txt
python-dotenv
langchain
pinecone-client
openai
langchain-openai
pypdf
langchain_pinecone